refactor: fix PR remarks
