+++ src/app/features/date-range/services/controller/date-range.service.ts
@@ -2,31 +2,15 @@ import { Injectable } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { Observable, of, switchMap } from 'rxjs';
 import { DateRangeComponent } from '../../components/date-range/date-range.component';
-import { UnavailableDates } from '../../models';
-import { PersonalInfoService } from '@app/shared/services/personal-info/personal-info.service';
-import { AuthState } from '@app/auth/store';
-import { Select } from '@ngxs/store';
+import { DateRangeData, UnavailableDates } from '../../models';
 
 @Injectable({
   providedIn: 'root',
 })
 export class DateRangeService {
-  @Select(AuthState.hasUserPesonalData) hasUserPesonalData$!: Observable<boolean>;
+  constructor(private dialog: MatDialog) {}
 
-  constructor(private dialog: MatDialog, private personalInfoService: PersonalInfoService) {}
-
-  openDateRangeModal(
-    unavailableDates: UnavailableDates[],
-    equipmentId?: number,
-    maxRentalPeriod?: number,
-  ): Observable<UnavailableDates | null> {
-    const dateRangeData = {
-      headerText: 'Период аренды',
-      buttonText: 'Подтвердить период аренды',
-      maxRentalPeriod,
-      unavailableDates,
-    };
-
+  openDateRangeModal(dateRangeData: DateRangeData): Observable<UnavailableDates | null> {
     return this.dialog
       .open(DateRangeComponent, {
         data: dateRangeData,
@@ -38,22 +22,8 @@ export class DateRangeService {
       .afterClosed()
       .pipe(
         switchMap((period: UnavailableDates | null) => {
-          return this.addPersonalInfo(period);
+          return period ? of(period) : of(null);
         }),
       );
-  }
-
-  private addPersonalInfo(period: UnavailableDates | null): Observable<UnavailableDates | null> {
-    if (!period) return of(null);
-
-    return this.hasUserPesonalData$.pipe(
-      switchMap((isPersonalData) => {
-        return !isPersonalData ? this.personalInfoService.openPersonalInfoModal() : of(null);
-      }),
-      switchMap(() => this.hasUserPesonalData$),
-      switchMap((isPersonalData) => {
-        return isPersonalData ? of(period) : of(null);
-      }),
-    );
   }
 }
