+++ src/app/shared/services/info/info.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { InfoModalComponent } from '@app/shared/components/info-modal/info-modal.component';
 import { InfoData } from '@app/shared/models';
+import { Observable, of, switchMap } from 'rxjs';
 
 @Injectable({
   providedIn: 'root',
@@ -9,13 +10,20 @@ import { InfoData } from '@app/shared/models';
 export class InfoService {
   constructor(private readonly dialog: MatDialog) {}
 
-  openInfoModal(infoData: InfoData) {
-    this.dialog.open(InfoModalComponent, {
-      data: infoData,
-      width: '390px',
-      maxWidth: '390px',
-      autoFocus: false,
-      position: { bottom: '0px' },
-    });
+  openInfoModal(infoData: InfoData): Observable<boolean | null> {
+    return this.dialog
+      .open(InfoModalComponent, {
+        data: infoData,
+        width: '390px',
+        maxWidth: '390px',
+        autoFocus: false,
+        position: { bottom: '0px' },
+      })
+      .afterClosed()
+      .pipe(
+        switchMap((buttonValue: boolean | undefined) => {
+          return typeof buttonValue === 'boolean' ? of(buttonValue) : of(null);
+        }),
+      );
   }
 }
